rename the 2 mixins to better differentiate them
rename the 2 mixins to better differentiate them
