+++ src/mixins/indexDataFetcher.js
