+++ src/mixins/indexDataGetter.js
